+++ cardano-node/ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Next version
 
+- Removed `cardano-node' as a dependency from `cardano-tracer'. This necessitated moving `NodeInfo` from
+  `cardano-tracer:Cardano.Node.Startup`to `trace-dispatcher:Cardano.Logging.Types.NodeInfo`, and `NodePeers` from
+  `cardano-node:Cardano.Node.Tracing.Peers` to `trace-dispatcher:Cardano.Logging.Types.NodePeers`.
+
 - Optionally support lightweight checkpointing.
 
   This PR adds optional support for *lightweight checkpointing*. Concretely, a file can contain a list of checkpoints (each consisting of a block number and a corresponding block hash). When validating a header/block with a block number with a corresponding checkpoint, we consider the header/block to be invalid if their actual hash does not coincide with the hash from the checkpoint.